Renamed Files + Removed Console Logs
