+++ __test__/logInForm.test.tsx
@@ -3,7 +3,7 @@ import { render, screen } from '@testing-library/react';
 import '@testing-library/jest-dom';
 import { MantineProvider } from '@mantine/core';
 import userEvent from '@testing-library/user-event';
-import LoginForm from '../src/components/Authorization/loginForm.client';
+import LoginForm from '../src/components/Authorization/LoginForm/LoginForm';
 
 const mockedSignIn = require('@aws-amplify/auth').signIn;
