+++ app/page.tsx
@@ -3,7 +3,7 @@
 import React from 'react';
 import { Box } from '@mantine/core';
 import Image from 'next/image';
-import LoginForm from '@/src/components/Authorization/loginForm.client';
+import LoginForm from '@/src/components/Authorization/LoginForm/LoginForm';
 
 const containerStyling = {
   display: 'flex',
@@ -12,8 +12,6 @@ const containerStyling = {
   height: '100vh',
 };
 export default function MainPage() {
-  console.log('in login');
-
   return (
     <div style={containerStyling}>
       <Box>
