+++ src/components/Home/CreatePostDrawer/CreatePostDrawer.tsx
@@ -86,9 +86,7 @@ export function CreatePostDrawer({ opened, onClose, onPostCreated }: CreatePostD
             type="button"
             onClick={() => {
               formik.validateForm().then((errors) => {
-                console.log(errors); // For logging
                 if (Object.keys(errors).length === 0 && !loading) {
-                  // No errors, form is valid so we submit
                   formik.submitForm();
                 } else {
                   notifications.show({
