+++ src/components/Marketplace/CreateListingDrawer/createListingValidation.ts
