+++ src/components/Marketplace/MarketplaceFeed/MarketplaceFeed.tsx
@@ -1,8 +1,8 @@
 import { useState } from 'react';
 import { Group, Loader, SimpleGrid, Text } from '@mantine/core';
 import { ItemForSale } from '@/src/types/types';
-import { MarketplaceCard } from '@/src/components/Marketplace/MarketplaceCard';
-import { ViewListingModal } from '@/src/components/Marketplace/ViewListingModal';
+import { MarketplaceCard } from '@/src/components/Marketplace/MarketplaceCard/MarketplaceCard';
+import { ViewListingModal } from '@/src/components/Marketplace/ViewListingModal/ViewListingModal';
 import { useFetchListings, useUserListingSaves } from '@/src/hooks/marketplaceCustomHooks';
 import { filterAndSortListings } from '@/src/components/utils/marketplaceUtils';
 import { useCurrentUser } from '@/src/hooks/usersCustomHooks';
