+++ src/components/Profile/AccountSettingsModal/AccountSettingsModal.tsx
@@ -123,11 +123,8 @@ export function AccountSettingsModal({ opened, onClose }: AccountSettingsModalPr
                 type="button"
                 data-testid="submit-btn"
                 onClick={() => {
-                  console.log(formik.values); // For logging
                   formik.validateForm().then((errors) => {
-                    console.log(errors); // For logging
                     if (Object.keys(errors).length === 0 && !loading) {
-                      // No errors, form is valid so we submit
                       formik.submitForm();
                     } else {
                       notifications.show({
